fix: move some methods to CreateIdentityViewModel
